+++ apiEstudo/Application/Arguments/Base/BaseInputIdentityUpdate.cs
@@ -2,10 +2,10 @@
 {
     public class BaseInputIdentityUpdate<TInputUpdate> where TInputUpdate : BaseInputUpdate<TInputUpdate>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public virtual TInputUpdate? InputUpdate { get; set; }
 
-        public BaseInputIdentityUpdate(int id, TInputUpdate? inputUpdate)
+        public BaseInputIdentityUpdate(long id, TInputUpdate? inputUpdate)
         {
             Id = id;
             InputUpdate = inputUpdate;
@@ -14,6 +14,6 @@
 
     public class BaseInputIdentityUpdate_0 : BaseInputIdentityUpdate<BaseInputUpdate_0>
     {
-        public BaseInputIdentityUpdate_0(int id, BaseInputUpdate_0 inputUpdate) : base(id, inputUpdate) { }
+        public BaseInputIdentityUpdate_0(long id, BaseInputUpdate_0 inputUpdate) : base(id, inputUpdate) { }
     }
 }
