+++ apiEstudo/Application/Arguments/Product/InputIdentityUpdateProduct.cs
@@ -2,7 +2,7 @@
 {
     public class InputIdentityUpdateProduct : BaseInputIdentityUpdate<InputUpdateProduct>
     {
-        public InputIdentityUpdateProduct(int id, InputUpdateProduct inputUpdate) : base(id, inputUpdate)
+        public InputIdentityUpdateProduct(long id, InputUpdateProduct inputUpdate) : base(id, inputUpdate)
         {
         }
     }
