+++ apiEstudo/Application/Arguments/ShoppingItem/InputCreateShoppingItem.cs
@@ -2,7 +2,7 @@
 {
     public class InputCreateShoppingItem : BaseInputCreate<InputCreateShoppingItem>
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public double Quantity { get; set; }
