+++ apiEstudo/Application/Arguments/ShoppingItem/InputUpdateShoppingItem.cs
@@ -2,7 +2,7 @@
 {
     public class InputUpdateShoppingItem : BaseInputUpdate<InputUpdateShoppingItem>
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public double Quantity { get; set; }
     }
 }
