+++ apiEstudo/Application/ServicesInterfaces/IShoppingService.cs
@@ -4,7 +4,7 @@ namespace apiEstudo.Application.ServicesInterfaces
 {
     public interface IShoppingService : IBaseService<InputCreateShopping, InputCreateShoppingComplete, InputInternalCreateShopping, InputUpdateShopping, InputIdentityUpdateShopping, InputIdentityDeleteShopping, OutputShopping>
     {
-        public int UpdateShippingStatusApprove(InputApproveShippingStatus inputApproveShippingStatus);
-        public int UpdateShippingStatusCancel(InputCancelShippingStatus inputCancelShippingStatus);
+        public long UpdateShippingStatusApprove(InputApproveShippingStatus inputApproveShippingStatus);
+        public long UpdateShippingStatusCancel(InputCancelShippingStatus inputCancelShippingStatus);
     }
 }
